Fixed a memory leak in set_text

+++ common/dwr.h
@@ -97,7 +97,7 @@ extern "C" {
  */
 uint64_t dwr_randomize(const char* input_file, uint64_t seed, char *flags,
         const char *sprite_name, const char* output_dir);
-size_t set_text(dw_rom *rom, const size_t address, char *text);
+uint16_t set_text(dw_rom *rom, const size_t address, char *text);
 
 #ifdef __cplusplus
 }
